fix refactor issues
