+++ pkg/common/token.go
@@ -5,7 +5,7 @@ import (
 	"net/url"
 	"os"
 
-	"github.com/go-resty/resty/v2"
+	resty "github.com/go-resty/resty/v2"
 )
 
 type Cred struct {
